feature offline mode. (#12)
*  - hover doesn't update on every run
  - fixes: https://github.com/go-flutter-desktop/go-flutter/issues/169
 - provide a `hover upgrade `command (like flutter upgrade)
 - every 2 days a check is made to see if the current project tags uses the last
   github release. if a new release is available print a message about
   `hover upgrade`
 - adds more debug message, 'bundling flutter', 'compiling
   go-flutter', 'running project',...
 - only print the 'channel beta' message if the user is on master,
  if the user was on stable the message was printed for no reason.

The behaviour of `hover run` and `hover build` is the same as before
when `--branch` is provided.
But there is a little twist, if the current version is a valid release
(available on github), `hover run` will not overwrite the 'go-flutter'
version. Example:

```
$ hover run --branch '@v0.20.0'
$ hover run # no upgrade trigged running on v0.20.0, at some point you
            # will get a `hover upgrade` notice

$ hover upgrade # user has to run to upgrade away from v0.20.0

$

$ hover run --branch '@master'
$ hover run # auto upgrade trigged! Not a real release.
            # If the user hasn't access to internet, fail silently and
            # keep using '@master'
```

To check every two days, a hidden file `./desktop/.last_goflutter_check`
is created on a project bases, the file contains a timestamps indicating
the last hover update check.

About the hover repo:
A empty go.mod file has been added to hover/assets/app/ this prevent adding
`go-flutter` as a go mod dependency. (because `go mod tidy` was peeking
listing hover/asset/options.go as a hover code file)

* chore: add last_goflutter_check into gitignore

* feat:run omit building flutter project or embedder

When using hover only for building plugin the `flutter build bundle`
slows down the process. Add a flag to omit `flutter build bundle`.

When making changes to the flutter package only, omit `go build
'go-flutter'` with the flag omit-embedder.

* doc: expose the branch hover flag

* fix runcmd and add observatory-port

* ecosystem consideration: prefixing files with 'go-' (#13)

* chore: add last_goflutter_check into gitignore

* ecosystem consideration: prefixing files with 'go-'

* add --path flag to all commands

* chore: fix indent

* Update init.go

* address the PR comments

* address PR comment

* move enginecache func to new versioncheck package

* updates the hover's gitignore file when migrating

adds the '.last_goflutter_check' entry to the go/.gitignore file if
needed

* internal/versioncheck: (chore) Renamed utility function